+++ js/chatbot/src/utils.ts
@@ -189,33 +189,38 @@ async function copy_to_clipboard(value: string) {
 
 export { marked };
 
-export const format_chat_for_sharing = async (chat: Array<
-	[string | FileData | null, string | FileData | null]
->) => {
-	let messages = await Promise.all(chat.map(async (message_pair) => {
-		return await Promise.all(message_pair
-			.map(async (message, i) => {
-				if (message === null) return "";
-				let speaker_emoji = i === 0 ? "👤" : "🤖";
-				let html_content = "";
-				if (typeof message === "string") {
-					html_content = message;
-				} else {
-					const file_url = await uploadToHuggingFace(message.data, "url");
-					if (message.mime_type?.includes("audio")) {
-						html_content = `<audio controls src="${file_url}"></audio>`;
-					} else if (message.mime_type?.includes("video")) {
-						html_content = file_url;
-					} else if (message.mime_type?.includes("image")) {
-						html_content = `<img src="${file_url}" />`;
+export const format_chat_for_sharing = async (
+	chat: Array<[string | FileData | null, string | FileData | null]>
+) => {
+	let messages = await Promise.all(
+		chat.map(async (message_pair) => {
+			return await Promise.all(
+				message_pair.map(async (message, i) => {
+					if (message === null) return "";
+					let speaker_emoji = i === 0 ? "👤" : "🤖";
+					let html_content = "";
+					if (typeof message === "string") {
+						html_content = message;
+					} else {
+						const file_url = await uploadToHuggingFace(message.data, "url");
+						if (message.mime_type?.includes("audio")) {
+							html_content = `<audio controls src="${file_url}"></audio>`;
+						} else if (message.mime_type?.includes("video")) {
+							html_content = file_url;
+						} else if (message.mime_type?.includes("image")) {
+							html_content = `<img src="${file_url}" />`;
+						}
 					}
-				}
-				return `${speaker_emoji}: ${html_content}`;
-			}))
-	}));
-	return messages.map(
-		(message_pair) => message_pair.join(
-			message_pair[0] !== "" && message_pair[1] !== "" ? "\n" : "")
-	).join("\n");
-}
-
+					return `${speaker_emoji}: ${html_content}`;
+				})
+			);
+		})
+	);
+	return messages
+		.map((message_pair) =>
+			message_pair.join(
+				message_pair[0] !== "" && message_pair[1] !== "" ? "\n" : ""
+			)
+		)
+		.join("\n");
+};